Refactor Workspace Settings page: split components into separate files for better organization and maintainability.
