+++ view/desktop/src/pages/WorkspaceSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { WorkspaceSettings } from "./WorkspaceSettingsPage";
